bemade_fsm 15.0.1.0.2 - add visits to FSM orders when missing
All sale orders containing a line that will produce an FSM task,
i.e. lines that are linked to a product of type service and with
service_tracking set to task_global_project, will be processed
to make sure that all such lines fall under a service visit. If
a visit already exists on the SO but lines are not covered by it,
the visit will be promoted to be the top item on the order before
confirmation. Otherwise, a default visit will be added to the top
of the order prior to proceeding with confirmation.