new module account_email_to_pdf
Since Odoo 18, emails coming in to an alias creating account moves
(vendor bills) get rejected if they do not contain an attachment that
can be read by the system. This means that sending a plain email receipt
with no attachment bounces, when it would be nice to have a vendor bill
with the message in the chatter as a minimum.

This module checks for attachments and injects one, in the form of a
simple pdf containing the email header and contents, if there was no
attachment to begin with.

This should enable sending of a simple email and not having it bounce
due to there being no attachment.