[ADD] New module stock_quant_reserved_fix (admin powers)
Adds a new module to allow admins to fix the stock quant reserved
quantity manually from the Physical Inventories list (or other
stock.quant list views).

This is necessary because the reserved quantity sometimes goes out of
whack with stock transfers cancelling or otherwise changing but not
updating the stock quant to unreserve the quantities they had previously
reserved. This is, of course, less ideal than fixing the issue at its
core, but it helps to have a tool for intervention when necessary.